Try adding python-version to key
Previous change didn't work as python 3.6 loaded the old cache so try
adding the python-version as part of the key.

I believe this will work first time, but probably not the second because
python 3.8 failed in "Test pytest", there was no pytest.

Add `echo PATH=$PATH` to see how PATH differs when working and non-working.

Add `which xxx` to see where python, pytest and pip are found when it
does work.